Update index.md
My relay shows a phantom voltage of about 75V when it is turned off and there is no load.  So I think the template relay state on boot needs to be set with a threshold of 90 or so (which is a guess, but it works for me).  

Also, the dynamic template relay state doesn't need to be optimistic based on this; I think it should be set based on the same lambda that determines its state on startup. 

This may have some side-effects if the mains voltage is less than 90 (relay would always show "off").  But, as it is now, my relay will always turn on on startup (since it will read an "on" voltage of 75, which is higher than the threshold of 10).  Another option would be to just keep the template switch optimistic and turn the template relay (and physical relay) off on boot, without trying to determine its actual startup state.